+++ src/main/java/com/backend/ClinicaOdontologica/dto/salida/OdontologoSalidaDto.java
@@ -2,7 +2,7 @@ package com.backend.ClinicaOdontologica.dto.salida;
 
 public class OdontologoSalidaDto {
 
-    private int id;
+    private Long id;
 
     private String matricula;
 
@@ -19,7 +19,7 @@ public class OdontologoSalidaDto {
     }
 
 
-    public OdontologoSalidaDto(int id, String matricula, String nombre, String apellido) {
+    public OdontologoSalidaDto(Long id, String matricula, String nombre, String apellido) {
         this.id = id;
         this.matricula = matricula;
         this.nombre = nombre;
@@ -27,11 +27,11 @@ public class OdontologoSalidaDto {
     }
 
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
