+++ src/main/java/com/backend/ClinicaOdontologica/dto/salida/PacienteSalidaDto.java
@@ -4,7 +4,7 @@ import java.time.LocalDate;
 
 public class PacienteSalidaDto {
 
-    private int id;
+    private Long id;
     private String nombre;
     private String apellido;
     private int dni;
@@ -16,7 +16,7 @@ public class PacienteSalidaDto {
     public PacienteSalidaDto() {
     }
 
-    public PacienteSalidaDto(int id, String nombre, String apellido, int dni, LocalDate fechaIngreso, DomicilioSalidaDto domicilioSalidaDto) {
+    public PacienteSalidaDto(Long id, String nombre, String apellido, int dni, LocalDate fechaIngreso, DomicilioSalidaDto domicilioSalidaDto) {
         this.id = id;
         this.nombre = nombre;
         this.apellido = apellido;
@@ -26,11 +26,11 @@ public class PacienteSalidaDto {
     }
 
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
